streamlit-mvp/pyproject.toml
[project]
name = "streamlit-mvp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.4.3",
    "google-genai>=1.13.0",
    "openai>=1.77.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.8",
]
